Use the new register_meta() functionality in conjunction with its sanitizing method with the framework meta.
git-svn-id: http://svn.locallylost.com/themes/hybrid-core/trunk@940 dba0f204-706d-4bc1-bc29-8b92e0485636

+++ admin/meta-box-post-seo.php
@@ -78,14 +78,10 @@ function hybrid_meta_box_post_save_seo( $post_id, $post ) {
 	if ( !isset( $_POST['hybrid-core-post-seo'] ) || !wp_verify_nonce( $_POST['hybrid-core-post-seo'], basename( __FILE__ ) ) )
 		return $post_id;
 
-	/* Check if the current user has permission to edit the post meta. */
-	if ( !current_user_can( 'edit_post_meta', $post_id ) )
-		return $post_id;
-
 	$meta = array(
-		'Title' => 	strip_tags( $_POST['hybrid-document-title'] ),
-		'Description' => 	strip_tags( $_POST['hybrid-meta-description'] ),
-		'Keywords' => 	strip_tags( $_POST['hybrid-meta-keywords'] )
+		'Title' => 	$_POST['hybrid-document-title'],
+		'Description' => 	$_POST['hybrid-meta-description'],
+		'Keywords' => 	$_POST['hybrid-meta-keywords']
 	);
 
 	foreach ( $meta as $meta_key => $new_meta_value ) {
@@ -94,11 +90,15 @@ function hybrid_meta_box_post_save_seo( $post_id, $post ) {
 		$meta_value = get_post_meta( $post_id, $meta_key, true );
 
 		/* If there is no new meta value but an old value exists, delete it. */
-		if ( '' == $new_meta_value && $meta_value )
+		if ( current_user_can( 'delete_post_meta', $post_id, $meta_key ) && '' == $new_meta_value && $meta_value )
 			delete_post_meta( $post_id, $meta_key, $meta_value );
 
+		/* If a new meta value was added and there was no previous value, add it. */
+		elseif ( current_user_can( 'add_post_meta', $post_id, $meta_key ) && $new_meta_value && '' == $meta_value )
+			add_post_meta( $post_id, $meta_key, $new_meta_value, true );
+
 		/* If the new meta value does not match the old value, update it. */
-		elseif ( $new_meta_value && $new_meta_value != $meta_value )
+		elseif ( current_user_can( 'edit_post_meta', $post_id, $meta_key ) && $new_meta_value && $new_meta_value != $meta_value )
 			update_post_meta( $post_id, $meta_key, $new_meta_value );
 	}
 }
