+++ admin/meta-box-post-template.php
@@ -80,16 +80,12 @@ function hybrid_meta_box_post_save_template( $post_id, $post ) {
 	if ( !isset( $_POST['hybrid-core-post-meta-box-template'] ) || !wp_verify_nonce( $_POST['hybrid-core-post-meta-box-template'], basename( __FILE__ ) ) )
 		return $post_id;
 
-	/* Check if the current user has permission to edit the post meta. */
-	if ( !current_user_can( 'edit_post_meta', $post_id ) )
-		return $post_id;
-
 	/* Return here if the template is not set. There's a chance it won't be if the post type doesn't have any templates. */
 	if ( !isset( $_POST['hybrid-post-template'] ) )
 		return $post_id;
 
 	/* Get the posted meta value. */
-	$new_meta_value = strip_tags( $_POST['hybrid-post-template'] );
+	$new_meta_value = $_POST['hybrid-post-template'];
 
 	/* Set the $meta_key variable based off the post type name. */
 	$meta_key = "_wp_{$post->post_type}_template";
@@ -98,11 +94,15 @@ function hybrid_meta_box_post_save_template( $post_id, $post ) {
 	$meta_value = get_post_meta( $post_id, $meta_key, true );
 
 	/* If there is no new meta value but an old value exists, delete it. */
-	if ( '' == $new_meta_value && $meta_value )
+	if ( current_user_can( 'delete_post_meta', $post_id ) && '' == $new_meta_value && $meta_value )
 		delete_post_meta( $post_id, $meta_key, $meta_value );
 
+	/* If a new meta value was added and there was no previous value, add it. */
+	elseif ( current_user_can( 'add_post_meta', $post_id, $meta_key ) && $new_meta_value && '' == $meta_value )
+		add_post_meta( $post_id, $meta_key, $new_meta_value, true );
+
 	/* If the new meta value does not match the old value, update it. */
-	elseif ( $new_meta_value && $new_meta_value != $meta_value )
+	elseif ( current_user_can( 'edit_post_meta', $post_id ) && $new_meta_value && $new_meta_value != $meta_value )
 		update_post_meta( $post_id, $meta_key, $new_meta_value );
 }
